services/clip-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.1
transformers==4.35.2
pillow==10.1.0
python-multipart==0.0.6
numpy<2.0.0 